More Dagger tweaking; remove need to subclass DaggerApplication
